requirements.txt
thrift==0.8.0
Twisted==12.1.0
zope.interface==4.0.1
mock
